Fix unintended attempt of reloading inactive httpd.service
Currently, httpd.service is reloaded by /etc/logrotate.d/httpd
regardless of the startup status of the httpd.service. At this time,
if the reload is executed when the httpd.service is not running, the
following message will be displayed in the log such as systemd journal
and /var/log/messages:

    systemd[1]=unit httpd.service cannot be reloaded because it is inactive

To prevent this, add checking the startup state of httpd to the
rotation script. Make sure that the httpd service reload is performed
only when the httpd state is in the startup state.